refactor(cognition): apply_act extracted from act_observe into apply.rs (pure code-motion)
Moved apply_act (stays pub, re-exported via `pub use apply::apply_act`), its private helper
is_long_running, and the PROPRIOCEPTION_RECALL_SALIENCE const verbatim into a new apply.rs.
Cross-module calls resolve via `use super::recency::{...}`, `use super::perception::{...}`, and
`use super::now_ms`. The one self-contained is_long_running test moved with the function into
apply.rs's own `#[cfg(test)] mod tests`. One doc-link path (ActingBody) rewritten super::→crate::
to stay resolvable from the new location. Behavior-identical; no external caller changes.
mod.rs 2651 → 2159 lines; apply.rs is 401 lines. act_observe tests green: 28 passed, 0 failed.
